+++ test/__mocks/AssetManager.sol
@@ -7,27 +7,27 @@ import { IAssetManager } from "src/interfaces/IAssetManager.sol";
 import { ReservoirPair } from "src/ReservoirPair.sol";
 
 contract AssetManager is IAssetManager {
-    mapping(ReservoirPair => mapping(ERC20 => uint104)) public getBalance;
+    mapping(ReservoirPair => mapping(ERC20 => uint256)) public getBalance;
 
     function adjustManagement(ReservoirPair aPair, int256 aToken0Amount, int256 aToken1Amount) external {
-        require(aToken0Amount != type(int224).min && aToken1Amount != type(int224).min, "AM: OVERFLOW");
+        require(aToken0Amount != type(int256).min && aToken1Amount != type(int256).min, "AM: OVERFLOW");
 
         if (aToken0Amount >= 0) {
-            uint104 lAbs = uint104(uint256(int256(aToken0Amount)));
+            uint256 lAbs = uint256(int256(aToken0Amount));
 
             getBalance[aPair][aPair.token0()] += lAbs;
         } else {
-            uint104 lAbs = uint104(uint256(int256(-aToken0Amount)));
+            uint256 lAbs = uint256(int256(-aToken0Amount));
 
             aPair.token0().approve(address(aPair), lAbs);
             getBalance[aPair][aPair.token0()] -= lAbs;
         }
         if (aToken1Amount >= 0) {
-            uint104 lAbs = uint104(uint256(int256(aToken1Amount)));
+            uint256 lAbs = uint256(int256(aToken1Amount));
 
             getBalance[aPair][aPair.token1()] += lAbs;
         } else {
-            uint104 lAbs = uint104(uint256(int256(-aToken1Amount)));
+            uint256 lAbs = uint256(int256(-aToken1Amount));
 
             aPair.token1().approve(address(aPair), lAbs);
             getBalance[aPair][aPair.token1()] -= lAbs;
@@ -36,7 +36,7 @@ contract AssetManager is IAssetManager {
         aPair.adjustManagement(aToken0Amount, aToken1Amount);
     }
 
-    function adjustBalance(ReservoirPair aOwner, ERC20 aToken, uint104 aNewAmount) external {
+    function adjustBalance(ReservoirPair aOwner, ERC20 aToken, uint256 aNewAmount) external {
         getBalance[aOwner][aToken] = aNewAmount;
     }
